Move format ratings into common util
