+++ client/app/intake/reducers/appeal.js
@@ -2,7 +2,8 @@ import { ACTIONS, REQUEST_STATE } from '../constants';
 import { FORM_TYPES } from '../../intakeCommon/constants';
 import { update } from '../../util/ReducerUtil';
 import { formatDateStr } from '../../util/DateUtil';
-import { getReceiptDateError, getPageError, formatRatings, formatRelationships } from '../util';
+import { getReceiptDateError, getPageError, formatRelationships } from '../util';
+import { formatRatings } from '../../intakeCommon/util';
 import _ from 'lodash';
 
 const getDocketTypeError = (responseErrorCodes) => (
